flask/requirements.txt
gunicorn==20.0.4
Flask==1.1.2
python-dateutil==2.8.2